requirements.txt
streamlit
pandas
openpyxl
requests
beautifulsoup4
pytesseract
pdf2image
Pillow
lxml
PyMuPDF
